build plugin using ng-packager
