+++ packages/plugin-angular/.gitignore
@@ -1,1 +1,7 @@
+# See https://docs.github.com/get-started/getting-started-with-git/ignoring-files for more about ignoring files.
+
+# Generated bundles for e2e testing
 bugsnag-plugin-angular-*.tgz
+
+# Miscellaneous
+.angular
