+++ packages/plugin-angular/package.json
@@ -2,13 +2,6 @@
   "name": "@bugsnag/plugin-angular",
   "version": "8.1.2",
   "description": "Angular integration for bugsnag-js",
-  "main": "dist/esm5/index.js",
-  "browser": "dist/esm5/index.js",
-  "module": "dist/esm5/index.js",
-  "es2015": "dist/esm2015/index.js",
-  "esm5": "dist/esm5/index.js",
-  "types": "dist/esm5/index.d.ts",
-  "typings": "dist/esm5/index.d.ts",
   "homepage": "https://www.bugsnag.com/",
   "repository": {
     "type": "git",
@@ -20,25 +13,35 @@
   "files": [
     "dist"
   ],
+  "type": "module",
+  "module": "dist/fesm2022/bugsnag-plugin-angular.mjs",
+  "types": "dist/index.d.ts",
   "scripts": {
     "clean": "rm -fr dist && mkdir dist",
-    "build": "npm run clean && npm run build:esm2015 && npm run build:esm5",
-    "build:esm2015": "ngc -p tsconfig.json",
-    "build:esm5": "ngc -p tsconfig.esm5.json",
+    "build": "ng build",
     "test:types": "npm run build"
   },
   "author": "Bugsnag",
   "license": "MIT",
-  "devDependencies": {
-    "@angular/cli": "^17.0.0",
-    "@angular/compiler": "^17.0.0",
-    "@angular/compiler-cli": "^17.0.0",
-    "@angular/core": "^17.0.0",
-    "@bugsnag/js": "^8.1.2",
-    "rxjs": "^7.4.0",
-    "zone.js": "~0.14.0"
-  },
   "peerDependencies": {
+    "@angular/common": ">= 12",
+    "@angular/core": ">= 12",
     "@bugsnag/js": "^8.0.0"
-  }
+  },
+  "devDependencies": {
+    "@angular/cli": "^19.0.5",
+    "@angular/common": "^19.0.0",
+    "@angular/compiler-cli": "^19.0.0",
+    "@angular/core": "^19.0.0",
+    "@angular-devkit/build-angular": "^19.0.5",
+    "@bugsnag/js": "^8.0.0",
+    "ng-packagr": "^19.0.0",
+    "rxjs": "~7.8.0",
+    "typescript": "~5.6.2",
+    "zone.js": "~0.15.0"
+  },
+  "dependencies": {
+    "tslib": "^2.3.0"
+  },
+  "sideEffects": false
 }
