+++ packages/plugin-angular/src/lib/bugsnag-error-handler.ts
@@ -1,0 +1,40 @@
+import { ErrorHandler, Injectable } from '@angular/core'
+import Bugsnag, { Client } from '@bugsnag/js'
+
+type BugsnagWithInternals = typeof Bugsnag & {
+  _client: Client
+}
+
+@Injectable()
+class BugsnagErrorHandler implements ErrorHandler {
+  public bugsnagClient: Client;
+
+  constructor (client?: Client) {
+    if (client) {
+      this.bugsnagClient = client
+    } else {
+      this.bugsnagClient = (Bugsnag as BugsnagWithInternals)._client
+    }
+  }
+
+  public handleError (error: any): void {
+    const handledState = {
+      severity: 'error',
+      severityReason: { type: 'unhandledException' },
+      unhandled: true
+    }
+
+    const event = this.bugsnagClient.Event.create(
+      error,
+      true,
+      handledState,
+      'angular error handler',
+      1
+    )
+
+    this.bugsnagClient._notify(event)
+    ErrorHandler.prototype.handleError.call(this, error)
+  }
+}
+
+export default BugsnagErrorHandler
