--- packages/plugin-angular/tsconfig.esm5.json
@@ -1,7 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "target": "ES5",
-    "outDir": "./dist/esm5"
-  }
-}
